# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: OPES School Software API
  description: >-
    Read-only integration surface over the modular monolith. REST is a thin
    adapter over the same module internals as the Livewire screens
    (00-core 6.1); it enforces the same permission gates and never exposes a
    write path in v1.
  version: 1.0.0
servers:
  - url: /api
security:
  - bearerToken: []
paths:
...
  /v1/me/announcements:
    get:
      operationId: listAnnouncements
      summary: School announcements
      description: >-
        Row 26, granted on any valid link, so not child-scoped. A guardian whose
        every link has expired gets 403 - 7.5's historic-access rule: an
        announcement is about a school you are currently part of. Scoped by
        PARTICIPATION rather than by re-deriving `announcement_recipients`
        scopes, because the office's scope is resolved into participants at send
        time and re-deriving it here would be a second, drifting answer to who
        was addressed.
      responses:
        '200':
          description: Announcements, newest first.
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /v1/me/notifications:
    get:
      operationId: listNotifications
      summary: This user's notifications
      description: >-
        Not matrix territory - `notifications.user_id` IS the scope. `meta.unread`
        is the badge count.
      parameters:
        - $ref: '#/components/parameters/perPage'
      responses:
        '200':
          description: Notifications, newest first, with an unread count.
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /v1/me/notifications/{notification}/read:
    post:
      operationId: markNotificationRead
      summary: Mark one notification read
      description: >-
        Ownership is checked by `Notifications\Actions\MarkNotificationRead` -
        one owner check, in the module that owns notifications. Another user's
        row answers 404, not 403: whether a notification id exists is not a
        guardian's business. Requires `portal.write`.
      parameters:
        - name: notification
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Marked read (idempotent). }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /v1/me/notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      summary: Mark every notification read
      description: Own rows only. Returns how many were changed. Requires `portal.write`.
      responses:
        '200': { description: Count of rows marked read. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /v1/me/threads:
    get:
      operationId: listThreads
      summary: Message threads this user participates in
      description: >-
        Authorized by PARTICIPATION, not by the matrix, and deliberately: a
        teacher who messaged a parent about one child would otherwise become
        unreachable the day that link expired, mid-conversation. Announcements
        are excluded - they have their own endpoint and their own rendering.
        Unread counts come from a message-id watermark, not a timestamp, because
        MySQL DATETIME is second-granular and a reply landing in the same second
        as a read stamp would silently read as already read.
      responses:
        '200':
          description: Threads, newest first, with unread counts.
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /v1/me/threads/{thread}/messages:
    get:
      operationId: listThreadMessages
      summary: Messages in one thread
      description: >-
        The participation check is repeated here rather than inferred from the
        list - a client may hold a thread id from any source, and the list is a
        convenience, never a control. A thread this user is not in answers 404.
      parameters:
        - name: thread
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Messages, oldest first. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    post:
      operationId: sendThreadMessage
      summary: Post a reply
      description: >-
        Body at most 4 000 characters (spec 5). Goes through
        `Communication\Actions\Messaging\PostMessage`, which re-checks
        participation inside its own transaction - the check on the way in is
        defence in depth, not the control. Requires `portal.write` and honours
        `Idempotency-Key`.
      parameters:
        - name: thread
          in: path
          required: true
          schema: { type: integer }
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
      responses:
        '201': { description: The posted message. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: Sanctum personal access token.
  parameters:
...
    perPage:
      name: per_page
      in: query
      description: Rows per page, 1-100.
      schema: { type: integer, default: 25, maximum: 100 }
  responses:
    Unauthenticated:
      description: Missing or invalid token.
    Forbidden:
      description: >-
        The user lacks the required permission, or the token lacks the
        required ability.
    NotFound:
      description: No such record.
    ValidationFailed:
      description: >-
        The request body failed validation. On the sign-in endpoint this is
        also the single, deliberately uninformative answer to every credential
        failure.
    TooManyRequests:
      description: >-
        A rate limit is exhausted - the 60/min `api` budget, the 5/min
        `auth-mobile` sign-in budget, or the 120/min `api-portal` device budget.
    Conflict:
      description: >-
        The same `Idempotency-Key` was replayed with a DIFFERENT body. The
        original response is not returned, because the caller evidently meant
        something else and silently answering the old request would hide that.
...